+++ src/Classifiers/MultiLayerPerceptron.php
@@ -314,7 +314,7 @@ class MultiLayerPerceptron implements Online, Probabilistic, Verbose, Persistabl
             return;
         }
 
-        !isset($this->logger) ?: $this->logger->info('Training started');
+        if ($this->logger) $this->logger->info('Training started');
 
         $n = $dataset->numRows();
 
@@ -347,7 +347,7 @@ class MultiLayerPerceptron implements Online, Probabilistic, Verbose, Persistabl
                 $bestSnapshot = Snapshot::take($this->network);
             }
 
-            !isset($this->logger) ?: $this->logger->info("Epoch $epoch"
+            if ($this->logger) $this->logger->info("Epoch $epoch"
                 . " completed, loss: $loss, score: $score");
 
             if ($score === $max) {
@@ -376,12 +376,12 @@ class MultiLayerPerceptron implements Online, Probabilistic, Verbose, Persistabl
             if (isset($bestSnapshot)) {
                 $this->network->restore($bestSnapshot);
 
-                !isset($this->logger) ?: $this->logger->info('Network restored'
+                if ($this->logger) $this->logger->info('Network restored'
                     . ' from previous snapshot');
             }
         }
 
-        !isset($this->logger) ?: $this->logger->info("Training complete");
+        if ($this->logger) $this->logger->info("Training complete");
     }
 
     /**