+++ src/Regressors/Adaline.php
@@ -230,7 +230,7 @@ class Adaline implements Online, Verbose, Persistable
             return;
         }
 
-        !isset($this->logger) ?: $this->logger->info('Training started');
+        if ($this->logger) $this->logger->info('Training started');
 
         $n = $dataset->numRows();
         
@@ -255,7 +255,7 @@ class Adaline implements Online, Verbose, Persistable
                 $bestSnapshot = Snapshot::take($this->network);
             }
             
-            !isset($this->logger) ?: $this->logger->info("Epoch $epoch"
+            if ($this->logger) $this->logger->info("Epoch $epoch"
                 . " completed, loss: $loss");
 
             if (abs($previous - $loss) < $this->minChange) {
@@ -269,12 +269,12 @@ class Adaline implements Online, Verbose, Persistable
             if (isset($bestSnapshot)) {
                 $this->network->restore($bestSnapshot);
 
-                !isset($this->logger) ?: $this->logger->info('Network restored'
+                if ($this->logger) $this->logger->info('Network restored'
                     . ' from previous snapshot');
             }
         }
 
-        !isset($this->logger) ?: $this->logger->info("Training complete");
+        if ($this->logger) $this->logger->info("Training complete");
     }
 
     /**