+++ src/Regressors/MLPRegressor.php
@@ -305,7 +305,7 @@ class MLPRegressor implements Online, Verbose, Persistable
             return;
         }
 
-        !isset($this->logger) ?: $this->logger->info('Training started');
+        if ($this->logger) $this->logger->info('Training started');
 
         $n = $dataset->numRows();
 
@@ -338,7 +338,7 @@ class MLPRegressor implements Online, Verbose, Persistable
                 $bestSnapshot = Snapshot::take($this->network);
             }
 
-            !isset($this->logger) ?: $this->logger->info("Epoch $epoch"
+            if ($this->logger) $this->logger->info("Epoch $epoch"
                 . " completed, loss: $loss");
 
             if ($score === $max) {
@@ -365,12 +365,12 @@ class MLPRegressor implements Online, Verbose, Persistable
             if (isset($bestSnapshot)) {
                 $this->network->restore($bestSnapshot);
 
-                !isset($this->logger) ?: $this->logger->info('Network restored'
+                if ($this->logger) $this->logger->info('Network restored'
                     . ' from previous snapshot');
             }
         }
 
-        !isset($this->logger) ?: $this->logger->info("Training complete");
+        if ($this->logger) $this->logger->info("Training complete");
     }
 
     /**